==> All_Image_Script/4_ Aggregate Dicom to png.ipynb <==
# 1. NIH image processing into aggregate folder 
import os
import shutil
import openpyxl
import csv

# Path to NIH metadata file


# Path to image source and destination folders
image_source = ('/srv/store/jkim/peds_cxr/')  # edit this. this is where images are located.
parent = ('/home/jkim/research/peds_cxr/') # edit this. this is the parent folder. 

source_folder = image_source + 'original_data/NIH_Peds/NIH-CXR-14/images/'
destination_folder = image_source + 'aggregate/'

metadata_file_path = parent + 'peds_cxr_metadata/raw_metadata/nih_raw_metadata_jk.csv'


# Read the image IDs from the CSV file
image_ids = []
with open(metadata_file_path, 'r') as file:
    reader = csv.reader(file)
    for row in reader:
        image_id = row[0]
        if image_id:  
            image_ids.append(image_id)

# Loop through the image IDs and copy the corresponding files
for image_id in image_ids:
    image_name = str(image_id)
    source = os.path.join(source_folder, image_name)
    destination = os.path.join(destination_folder, image_name)
    
    # Copy the file
    if os.path.exists(source):
        shutil.copy(source, destination)
    else:
        print(f'Image {image_name} not found in source folder.')

print('Finished copying images.')

# move all images into the 'images' subfolder 
# Specify the source folder which contains the .png files and the destination subfolder
source_folder = image_source + 'Aggregate/'
destination_subfolder = image_source + 'Aggregate/images'

# Ensure that the destination subfolder exists, if not, create it
if not os.path.exists(destination_subfolder):
    os.makedirs(destination_subfolder)

# Iterate through each file in the source folder
for file_name in os.listdir(source_folder):
    # Check if the file is a .png file
    if file_name.endswith('.png'):
        # Construct the full file path
        source = os.path.join(source_folder, file_name)
        destination = os.path.join(destination_subfolder, file_name)
        
        # Move the file to the destination subfolder
        shutil.move(source, destination)

print("Finished moving .png files.")
